Implement proof of concept full flow inserting set of unique player names from logs into table after parsing and aggregating

+++ pygw2agg/models/ei_output/player_defenses_stats.py
@@ -1,19 +1,20 @@
+from decimal import Decimal
 from pydantic import BaseModel
 
 
 class PlayerDefensesStats(BaseModel):
-    damageTaken: int
-    breakbarDamageTaken: int
-    blockedCount: int
-    evadedCount: int
-    missedCount: int
-    dodgeCount: int
-    invulnedCount: int
-    damageBarrier: int
-    interruptedCount: int
-    downCount: int
-    downDuration: int
-    deadCount: int
-    deadDuration: int
-    dcCount: int
-    dcDuration: int
+    damageTaken: Decimal
+    breakbarDamageTaken: Decimal
+    blockedCount: Decimal
+    evadedCount: Decimal
+    missedCount: Decimal
+    dodgeCount: Decimal
+    invulnedCount: Decimal
+    damageBarrier: Decimal
+    interruptedCount: Decimal
+    downCount: Decimal
+    downDuration: Decimal
+    deadCount: Decimal
+    deadDuration: Decimal
+    dcCount: Decimal
+    dcDuration: Decimal
