+++ pygw2agg/models/ei_output/player_dps_all_stats.py
@@ -1,18 +1,19 @@
+from decimal import Decimal
 from pydantic import BaseModel
 
 
 class PlayerDpsAllStats(BaseModel):
-    dps: int
-    damage: int
-    condiDps: int
-    condiDamage: int
-    powerDps: int
-    powerDamage: int
-    breakbarDamage: int
-    actorDps: int
-    actorDamage: int
-    actorCondiDps: int
-    actorCondiDamage: int
-    actorPowerDps: int
-    actorPowerDamage: int
-    actorBreakbarDamage: int
+    dps: Decimal
+    damage: Decimal
+    condiDps: Decimal
+    condiDamage: Decimal
+    powerDps: Decimal
+    powerDamage: Decimal
+    breakbarDamage: Decimal
+    actorDps: Decimal
+    actorDamage: Decimal
+    actorCondiDps: Decimal
+    actorCondiDamage: Decimal
+    actorPowerDps: Decimal
+    actorPowerDamage: Decimal
+    actorBreakbarDamage: Decimal
