+++ pygw2agg/models/ei_output/player_dps_target_stats.py
@@ -1,11 +1,12 @@
+from decimal import Decimal
 from pydantic import BaseModel
 
 
 class PlayerDpsTargetStats(BaseModel):
-    dps: int
-    damage: int
-    condiDps: int
-    condiDamage: int
-    powerDps: int
-    powerDamage: int
-    breakbarDamage: int
+    dps: Decimal
+    damage: Decimal
+    condiDps: Decimal
+    condiDamage: Decimal
+    powerDps: Decimal
+    powerDamage: Decimal
+    breakbarDamage: Decimal
